Extract schema variation merging into VariationMerger
SchemaBuilder exceeded the RuboCop class length limit. Its
build_merged_schema_from_variations method also exceeded the cyclomatic
complexity limit. Move the variation merging methods into a private
VariationMerger helper next to BuildContext. Split the object and array
merge branches into small methods so the main merge stays simple. The
array items merge that was duplicated in merge_multi now shares
merge_array_item_variations. No behavior change.
